requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
mysqlclient==2.2.8
packaging==26.0
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
